requirements.txt
flask
python-dotenv
pypandoc
pymupdf
pytesseract
pdfminer.six
pillow
requests
